+++ dragonchain/lib/database/redis.py
@@ -161,6 +161,11 @@ async def z_range_by_score_async(
     )
 
 
+async def srem_async(key: str, value: str) -> int:
+    await _set_redis_client_async_if_necessary()
+    return await async_redis_client.srem(key, value)
+
+
 async def get_async(key: str, *, decode: bool = True) -> Optional[str]:
     await _set_redis_client_async_if_necessary()
     return await async_redis_client.get(key, encoding="utf8" if decode else aioredis.util._NOTSET)
@@ -253,6 +258,11 @@ def lpush_sync(name: str, *values: str) -> int:
     return redis_client.lpush(name, *values)
 
 
+def sadd_sync(name: str, *values: str) -> int:
+    _set_redis_client_if_necessary()
+    return redis_client.sadd(name, *values)
+
+
 def rpush_sync(name: str, *values: str) -> int:
     _set_redis_client_if_necessary()
     return redis_client.rpush(name, *values)
@@ -310,11 +320,6 @@ def hget_sync(name: str, key: str, decode: bool = True) -> Optional[str]:
     _set_redis_client_if_necessary()
     response = redis_client.hget(name, key)
     return _decode_response(response, decode)
-
-
-def sadd_sync(name: str, *values: str) -> int:
-    _set_redis_client_if_necessary()
-    return redis_client.sadd(name, *values)
 
 
 def smembers_sync(name: str, decode: bool = True) -> set: